requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
requests
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
